+++ extensions/com_pungamail/components/com_pungamail/src/View/Confirm/HtmlView.php
@@ -9,6 +9,7 @@
 namespace Punga\Component\PungaMail\Site\View\Confirm;
 
 use Joomla\CMS\Factory;
+use Joomla\CMS\Language\Text;
 use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
 
 /** Punga Mail confirm view. */
@@ -20,6 +21,17 @@ final class HtmlView extends BaseHtmlView
 	public function display($tpl = null): void
 	{
 		$this->subscriber = $this->getModel()->getSubscriber();
+		$kind = Factory::getApplication()->getInput()->getCmd('kind');
+		$titleKey = 'COM_PUNGAMAIL_CONFIRM_INVALID_TITLE';
+
+		if ($this->subscriber !== null)
+		{
+			$titleKey = $kind === 'topics'
+				? 'COM_PUNGAMAIL_CONFIRM_TOPICS_TITLE'
+				: 'COM_PUNGAMAIL_CONFIRM_TITLE';
+		}
+
+		$this->setDocumentTitle(Text::_($titleKey));
 		parent::display($tpl);
 	}
 }