+++ extensions/com_pungamail/components/com_pungamail/src/View/Unsubscribe/HtmlView.php
@@ -8,7 +8,7 @@
 
 namespace Punga\Component\PungaMail\Site\View\Unsubscribe;
 
-use Joomla\CMS\Factory;
+use Joomla\CMS\Language\Text;
 use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
 
 /** Punga Mail unsubscribe view. */
@@ -20,6 +20,11 @@ final class HtmlView extends BaseHtmlView
 	public function display($tpl = null): void
 	{
 		$this->subscriber = $this->getModel()->getSubscriber();
+		$titleKey = $this->subscriber === null
+			? 'COM_PUNGAMAIL_UNSUBSCRIBE_INVALID_TITLE'
+			: 'COM_PUNGAMAIL_UNSUBSCRIBE_TITLE';
+
+		$this->setDocumentTitle(Text::_($titleKey));
 		parent::display($tpl);
 	}
 }